Try to get a sensible changelog again.
The so called 'breaking change' of picking the ZServer port was added in 1.3.1, not 1.3.0.
It was then backported to lots of branches, so probably not so breaking.
And then master apparently only had bug fixes, but there was a 1.3.x branch already,
so let's call that bugfixes and update the version in the next commit.
Oh well.

[ci skip]