Print ZServer and XMRPC URLs on server startup.
When starting the robot server, the ports for the ZServer and for the
XMLRPC-server can be configured to 0 in order to let the operating
system choose a random open port.

When the ports are randomly chosen, it is helpful that the server
prints out the chosen ports in form of URLs, so that a caller may read
the URLs in order to connect to the servers.